Break out timestamps code
