+++ packages/@dataform/sql/BUILD
@@ -1,5 +1,4 @@
 load("//tools:ts_library.bzl", "ts_library")
-load("//:version.bzl", "DF_VERSION")
 load("//packages:index.bzl", "pkg_bundle", "pkg_bundle_dts", "pkg_json", "pkg_npm_tar")
 
 package(default_visibility = ["//visibility:public"])
@@ -16,14 +15,14 @@ ts_library(
 pkg_json(
     name = "json",
     package_name = "@dataform/sql",
-    description = "General purpose SQL helper functions for building queries in JS across dialect.",
+    description = "General purpose SQL helper functions for building queries in JS across dialects.",
     layers = [
         "//:package.json",
         "//packages/@dataform:package.layer.json",
     ],
     main = "bundle.js",
     types = "bundle.d.ts",
-    version = DF_VERSION,
+    version = "0.1.0",
 )
 
 pkg_bundle(
