+++ sql/index.ts
@@ -5,13 +5,16 @@ import { JSONBuilder } from "df/sql/builders/json";
 import { ISelectOrBuilder, ISelectSchema } from "df/sql/builders/select";
 import { UnionBuilder } from "df/sql/builders/union";
 import { IWiths, WithBuilder } from "df/sql/builders/with";
-
-export type DurationUnit = "day" | "week" | "month" | "quarter" | "year";
+import { Timestamps } from "df/sql/timestamps";
 
 export type ISqlDialect = "standard" | "snowflake" | "postgres" | "mssql";
 
 export class Sql {
-  constructor(private readonly dialect: ISqlDialect = "standard") {}
+  public readonly timestamps: Timestamps;
+
+  constructor(private readonly dialect: ISqlDialect = "standard") {
+    this.timestamps = new Timestamps(dialect);
+  }
 
   public literal(value: string | number | null) {
     if (value === null) {
@@ -120,51 +123,6 @@ export class Sql {
 
   public safeDivide(numerator: string, denominator: string) {
     return `${numerator} / nullif(${denominator}, 0)`;
-  }
-
-  // Conversion functions.
-
-  public millisToTimestamp(timestampMillis: string) {
-    if (this.dialect === "snowflake") {
-      return `to_timestamp(${timestampMillis}, 3)`;
-    }
-    if (this.dialect === "postgres") {
-      return `timestamp 'epoch' + (${timestampMillis} / 1000) * interval '1 second'`;
-    }
-    return `timestamp_millis(${timestampMillis.toString()})`;
-  }
-
-  public timestampTruncate(timestamp: string, timestampUnit: DurationUnit) {
-    if (this.dialect === "snowflake") {
-      return `date_trunc(${timestampUnit}, ${timestamp})`;
-    }
-    if (this.dialect === "postgres") {
-      return `date_trunc('${timestampUnit}', ${timestamp})`;
-    }
-    return `timestamp_trunc(${timestamp}, ${timestampUnit})`;
-  }
-
-  public timestampToMillis(timestamp: string) {
-    if (this.dialect === "snowflake") {
-      return `date_part(epoch_milliseconds, ${timestamp})`;
-    }
-    if (this.dialect === "postgres") {
-      return `extract('epoch' from ${timestamp})::bigint * 1000`;
-    }
-    return `unix_millis(${timestamp})`;
-  }
-
-  public timestampCurrentUTC() {
-    if (this.dialect === "postgres") {
-      return "current_timestamp::timestamp";
-    }
-    if (this.dialect === "snowflake") {
-      return "convert_timezone('UTC', current_timestamp())::timestamp";
-    }
-    if (this.dialect === "mssql") {
-      return "CURRENT_TIMESTAMP";
-    }
-    return "current_timestamp()";
   }
 
   // Casting functions.
